+++ Shared/Model/Submissions.swift
@@ -35,7 +35,7 @@ struct SubmissionData: Codable {
 
 struct Attachment: Codable {
     let filename, type: String
-    let size: Int
+    let size: Double
     let url: URL
 }
